+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * [UITEN-338](https://folio-org.atlassian.net/browse/UITEN-338) Uniqueness check of location name/code is broken for \\ * ? ^
 * [UITEN-339](https://folio-org.atlassian.net/browse/UITEN-339) Invalidate react-query cache after location operations
 * [UITEN-335](https://folio-org.atlassian.net/browse/UITEN-335) Include universal read-only permissions
+* [UITEN-340](https://folio-org.atlassian.net/browse/UITEN-340) Use locale API to fetch and update tenant locale settings.
 
 ## [10.0.0](https://github.com/folio-org/ui-tenant-settings/tree/v10.0.0)(2025-03-12)